Fix Ticket #117

+++ datapack_development/sql/accounts.sql
@@ -4,7 +4,7 @@
 CREATE TABLE `accounts` (
   `login` VARCHAR(45) NOT NULL default '',
   `password` VARCHAR(45) ,
-  `lastactive` INT,
+  `lastactive` DECIMAL(20),
   `access_level` INT,
   `lastIP` VARCHAR(20),
   PRIMARY KEY (`login`)
